move as_regular_grammar, allow lexeme:bool kwarg
